WU-1.2/1.3: the run loop takes the tick lock, and the watchdog exists (I-34)
Two chassis guarantees that were specified long ago and enforced only on
`trdrbot tick` - which is to say, not on the path that runs unattended.

INV-7. `trdrbot run` called run_tick directly, so a launchd-driven run.sh
alongside a run loop would interleave two ticks freely: two tick-counter
read-modify-writes, two elfmem sessions on one SQLite file, two decide cycles
draining one inbox batch. There was a second, weaker lock here - a bare pid
file at the codebase's only RELATIVE path, no timestamp, never unlinked on
exit - which tick_lock supersedes in every respect. Deleted, along with the
dead `Path` annotation that made cli.py unresolvable to a type checker.

FM-26. tick.watchdog_seconds was configured and read by nothing, so a hung
LLM call stalled the run indefinitely, with no stale-lock signal to notice it
by either. Now bounded twice from one knob: asyncio.wait_for around the decide
call (the one thing that can hang for minutes - five retries per model, three
models, no bound of its own), and x4 that around the whole tick as the
backstop for a wedged MCP spawn or a stuck elfmem call. TimeoutError already
classifies TRANSIENT, so the existing handler journals it and retries the
batch next cycle.

watchdog_seconds raised 300 -> 600: measured decide cycles have reached 5:19
of wall clock (D-074), and a watchdog that kills legitimate work is worse than
no watchdog. Config property returns float so a test can use a sub-second
value without it truncating to zero.

A skipped tick now sleeps the OPEN interval rather than the closed one: a held
lock means another process is actively trading, so check back soon.

Verified by reverting the fix - both lock tests fail (assert [False] == [True],
assert [1] == []) and pass with it. The three _acquire_run_lock tests moved to
tick_lock rather than being deleted: they pin real properties (refuses a live
holder, breaks a stale lock, survives a corrupt file), the D-044 incident is
why those properties matter, and lock.py had no tests of its own before.

280 passed.

+++ src/trdrbot/config.py
@@ -85,8 +85,11 @@ class Config:
         return int(self.raw["inbox"]["max_retries"])
 
     @property
-    def watchdog_seconds(self) -> int:
-        return int(self.raw["tick"]["watchdog_seconds"])
+    def watchdog_seconds(self) -> float:
+        # float, not int: this bounds an `asyncio.wait_for`, and a test that
+        # wants a sub-second watchdog should not have it silently truncated
+        # to zero. Production values are whole seconds either way.
+        return float(self.raw["tick"]["watchdog_seconds"])
 
     def model_chain(self, role: str = "decide") -> list[str]:
         """Ordered fallback chain for a role (D-062).
